Labfiles/05-image-generation/Python/requirements.txt
python-dotenv
openai==1.10.0